Remove MetricName Interface and MetricNameFactory, change SeMetricName to MetricName
